move catalog/database code into data_mgmt module
