+++ app/src/commands.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 from typing import List, Optional
 import os
-from es_sfgtools.processing.pipeline.data_handler import DataHandler
+from es_sfgtools.data_mgmt.data_handler import DataHandler
 from es_sfgtools.utils.archive_pull import load_site_metadata,list_campaign_files
 from es_sfgtools.modeling.garpos_tools.load_utils import load_lib
 from .manifest import PipelineManifest
