+++ src/es_sfgtools/data_mgmt/catalog.py
