+++ src/es_sfgtools/data_mgmt/constants.py
