+++ src/es_sfgtools/pipelines/plotting.py
@@ -4,9 +4,9 @@ import numpy as np
 import datetime
 from typing import List
 
-from ..assets.tiledb import TDBGNSSArray   
-from ..assets.file_schemas import AssetEntry
-from ..pipeline.catalog import PreProcessCatalog
+from ..processing.assets.tiledb import TDBGNSSArray   
+from ..processing.assets.file_schemas import AssetEntry
+from ..data_mgmt.catalog import PreProcessCatalog
 
 UNIX_EPOCH = np.datetime64("1970-01-01T00:00:00Z")
 def to_timestamp(time:np.datetime64 | datetime.datetime) -> float:
